Machine Learning Models & Text Mining
You are hired by one of the leading news channel CNBE who wants to analyse recent elections. This survey was conducted on 1525 voters with 9 variables. You have to build a 
model, to predict which party a voter will vote for based on the given information, to create an exit poll that will help in predicting overall win and seats covered by a particular party. 
Dataset for Problem: Election_Data.xlsx

Text Mining : In this particular project, we are going to work on the inaugural corpora from the nltk in Python. We will be looking at the following speeches of the Presidents of the United States of America: 
1. President Franklin D. Roosevelt in 1941 
2. President John F. Kennedy in 1961 
3. President Richard Nixon in 1973