Market & Retail Analytics
The owner of the restaurant wants you to use this data to come up with a set of recommendations that can help his Café Chain increase his revenues. He has not been able to launch a loyalty program and is unable to provide you with a data set that has customer level information. But, he is able to provide you with a data set for POS (point of sale data) for one of his chains. 
• Executive summary :
 Tobacco contributes the highest sales.
 Sales are high during the night times compared to day.
 Customer purchases tobacco and food items together.
 Liquor Sales are high during night.
 Food items has a decline during the night where the day shows movement of food 
items