Capstone Project - Foreclosure of Loans
Defining Problem Statement: 
A Non-Banking Financial Company (NBFC) is a company registered under the Companies Act, 1956 engaged in the business of loans and advances, acquisition of shares/stocks/bonds/debentures/securities issued by Government or local authority or other marketable securities of a like nature, leasing, hire-purchase, insurance business, chit business but does not include any institution whose principal business is that of agriculture activity, industrial activity, purchase or sale of any goods (other than securities) or providing any services and sale/purchase/construction of immovable property. A non-banking institution which is a company and has principal business of receiving deposits under any scheme or arrangement in one lump sum or in installments by way of contributions or in any other manner, is also a non-banking financial company (Residuary non-banking company). 

Need of the study/project 
If the loan needs to be closed before the tenure, the lender may levy a prepayment penalty. The penalty is to cover the lost interest revenue from early closing of the loans. We need to analyze those customers who will be having the high probability in closing the loans taken from NBFC earlier than the tenor allotted.

Understanding business/social opportunity 
Financially foreclosure weaken the business of NBFC and they are prone to run mismatches in the short-term maturity. Our aim here is to predict those customers that can prone to be early payer of loans. We should present different services and products to the customers that will not encourage them to close the loans earlier than allotted tenor 
Foreclosure is a legal process in which a lender attempts to recover the balance of a loan from a borrower who has stopped making payments to the lender by forcing the sale of the asset used as the collateral for the loan.